source/sentence-bert-swedish-cased-cosine/requirements.txt
chromadb
fire
fitz
joblib
matplotlib
openai
pandas
python-dotenv
python-levenshtein
scikit-learn
seaborn
sentence-transformers
tomli
tomlkit
tqdm
PyMuPDF
